# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Employee Support Bot SaaS API
  version: 1.0.0
  description: |
    API documentation for the multi-tenant employee support bot backend.

    Intended production integrations:
    bot-facing `/api/v1/*` with `x-central-bot-secret`,
    optional partner `POST /api/v1/partner/tenants` with `x-tenant-provision-secret`,
    tenant-specific admin sessions for Platform / Company dashboards.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://your-platform-domain.example
    description: Production
...
security:
  - CentralBotSecret: []
paths:
...
  /api/v1/chat:
    post:
      tags: [Bot API]
      summary: Send an employee message to AI/RAG orchestration
      description: |
        Main API for central LINE bot integration.

        Authenticate with **`x-central-bot-secret`** or legacy **`x-api-key`** (no JSON-body auth secrets).
        For tenant-scoped **workflow** tokens when headers are impossible, use **`POST /api/v2/chat`** instead.

        If the user is not linked yet, send `company_code` together with the
        first message. The backend will link the user and answer in the same call.

        If the user is already linked, `external_user_id + channel` is enough.
      security:
        - CentralBotSecret: []
        - LegacyApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              afterRegistration:
                value:
                  external_user_id: legacy-user-uuid
                  channel: line
                  conversation_id: line-user-or-room-id
                  message: ลาป่วยได้กี่วัน
                  metadata:
                    line_user_id: Uxxxxxxxxxxxxxxxx
                    source: legacy-line-bot
              firstCallWithCompanyCode:
                value:
                  external_user_id: legacy-user-uuid
                  channel: line
                  company_code: ABC123
                  message: สวัสดิการประกันสุขภาพมีอะไรบ้าง
      responses:
        "200":
          description: Successful AI response, quota message, or bot-friendly registration prompt.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ChatSuccessResponse"
                  - $ref: "#/components/schemas/ChatNeedCompanyCodeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v2/chat:
    post:
      tags: [Bot API]
      summary: Simplified central bot chat (user_id, message, company_code)
      description: |
        LINE-oriented endpoint for central bot integrations.

        Authenticate with **`x-central-bot-secret`** (preferred), **or** send a
        tenant-scoped **`workflow_token`** in the JSON body when the HTTP client
        cannot set custom headers (same semantics as admin “Workflow HTTP” tokens).

        Request fields: `user_id`, `message`, and `company_code`. Backend resolves
        tenant, upserts user-to-tenant link where applicable, then runs the same
        AI orchestration pipeline as `/api/v1/chat`.
      security:
        - CentralBotSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatV2Request"
            examples:
              centralHeader:
                summary: With x-central-bot-secret
                value:
                  user_id: "111"
                  message: อยากคุยด้วย
                  company_code: ABC503
              workflowBody:
                summary: Headerless (workflow_token only)
                value:
                  workflow_token: wf_live_...
                  user_id: Uxxxxxxxxxxxxxxxx
                  message: อยากคุยด้วย
                  company_code: ABC503
      responses:
        "200":
          description: Successful AI response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User already linked to a different tenant (workflow_token path).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: TENANT_MISMATCH
                  message: This user is already linked to a different company.
        "404":
          description: Invalid or revoked company code (or invalid code for the workflow token tenant).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: INVALID_COMPANY_CODE
                  message: Invalid or revoked company code.
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    CentralBotSecret:
      type: apiKey
      in: header
      name: x-central-bot-secret
      description: Shared secret used by the trusted central LINE bot backend.
    LegacyApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Legacy tenant API key. Not recommended for central LINE bot integration.
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth browser session cookies. Actual cookie names depend on Supabase SSR configuration.
    TenantProvisionSecret:
      type: apiKey
      in: header
      name: x-tenant-provision-secret
      description: |
        Server-to-server secret for `POST /api/v1/partner/tenants`. Must match `TENANT_PROVISION_SECRET`
        in the deployment environment. Separate from `x-central-bot-secret`.
...
  responses:
    Unauthorized:
      description: Missing or invalid bot secret/API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: INVALID_API_KEY
              message: Invalid central bot secret.
...
    ValidationError:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  error:
                    type: string
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request body.
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: Internal server error.
...
  schemas:
...
    ChatRequest:
      type: object
      required: [external_user_id, message]
      properties:
        external_user_id:
          type: string
          maxLength: 255
          description: Stable user id from LINE or legacy system.
        channel:
          type: string
          default: api
          maxLength: 64
        message:
          type: string
          minLength: 1
          maxLength: 8000
        company_code:
          type: string
          maxLength: 64
          description: Optional. Used to link user to tenant if not linked yet.
        conversation_id:
          type: string
          maxLength: 255
        bot_id:
          type: string
          format: uuid
          deprecated: true
          description: Legacy field. Current runtime uses tenant AI settings.
        metadata:
          type: object
          additionalProperties: true
    ChatV2Request:
      type: object
      required: [user_id, message, company_code]
      properties:
        user_id:
          type: string
          maxLength: 255
          description: External user id from the central bot or legacy system.
        message:
          type: string
          minLength: 1
          maxLength: 8000
        company_code:
          type: string
          maxLength: 64
        workflow_token:
          type: string
          maxLength: 512
          description: |
            Optional. Tenant-scoped token when `x-central-bot-secret` cannot be sent as a header.
            Do not send `CENTRAL_BOT_SECRET` in this field.
    ChatSuccessResponse:
      type: object
      required:
        - success
        - reply
        - message_type
        - safety_level
        - conversation_id
        - sources
        - handoff_required
        - handoff
        - quota
      properties:
        success:
          type: boolean
          example: true
        reply:
          type: string
        message_type:
          type: string
          enum: [general, rag, mental_health, safety, crisis, out_of_scope, quota_exceeded]
        safety_level:
          type: string
          enum: [normal, medium, high, crisis]
        conversation_id:
          type: string
          format: uuid
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceDocument"
        handoff_required:
          type: boolean
        handoff:
          $ref: "#/components/schemas/Handoff"
        quota:
          $ref: "#/components/schemas/QuotaSnapshot"
        debug:
          type: object
          additionalProperties: true
    ChatNeedCompanyCodeResponse:
      type: object
      required: [success, reply, message_type, safety_level, conversation_id, sources, handoff_required, handoff, quota]
      properties:
        success:
          type: boolean
          example: true
        reply:
          type: string
          example: กรุณาลงทะเบียนด้วยรหัสบริษัทก่อนใช้งานครับ เพื่อให้ผมผูกบัญชี LINE ของคุณกับบริษัทที่ถูกต้อง
        message_type:
          type: string
          enum: [out_of_scope]
        safety_level:
          type: string
          enum: [normal]
        conversation_id:
          type: string
          example: ""
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceDocument"
        handoff_required:
          type: boolean
          example: false
        handoff:
          $ref: "#/components/schemas/Handoff"
        quota:
          $ref: "#/components/schemas/QuotaSnapshot"
    SourceDocument:
      type: object
      properties:
        document_name:
          type: string
        category:
          type: string
          nullable: true
          enum: [benefits, welfare, leave_policy, insurance, hr_faq, other, null]
        section:
          type: string
          nullable: true
        score:
          type: number
          format: float
    Handoff:
      type: object
      required: [enabled, url, button_text, message]
      properties:
        enabled:
          type: boolean
        url:
          type: string
          nullable: true
        button_text:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    QuotaSnapshot:
      type: object
      required: [plan, used, limit, remaining]
      properties:
        plan:
          type: string
          enum: [trial, pro]
        used:
          type: integer
        limit:
          type: integer
        remaining:
          type: integer
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string